# core/txpool/legacypool: refactor truncatePending